consolidate ga files
