+++ src/main/python/systemds/scuro/drsearch/multimodal_ga_optimizer.py
@@ -1,114 +1,59 @@
+# -------------------------------------------------------------
+#
+# Licensed to the Apache Software Foundation (ASF) under one
+# or more contributor license agreements.  See the NOTICE file
+# distributed with this work for additional information
+# regarding copyright ownership.  The ASF licenses this file
+# to you under the Apache License, Version 2.0 (the
+# "License"); you may not use this file except in compliance
+# with the License.  You may obtain a copy of the License at
+#
+#   http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing,
+# software distributed under the License is distributed on an
+# "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+# KIND, either express or implied.  See the License for the
+# specific language governing permissions and limitations
+# under the License.
+#
+# -------------------------------------------------------------
+
 from __future__ import annotations
 
 import copy
 import multiprocessing as mp
+import os
 import pickle
 import random
+import tempfile
 import time
+import traceback
+from concurrent.futures import FIRST_COMPLETED, ProcessPoolExecutor, wait
 from dataclasses import dataclass, field
-from concurrent.futures import ProcessPoolExecutor, wait, FIRST_COMPLETED
-from typing import Any, Dict, Generator, List, Optional, Tuple
+from itertools import chain
+from typing import Any, Dict, List, Optional, Tuple
+
+from deap import base, creator, tools
 
 from systemds.scuro.drsearch.operator_registry import Registry
 from systemds.scuro.drsearch.representation_dag import (
-    RepresentationDag,
     RepresentationDAGBuilder,
+    RepresentationDag,
 )
 from systemds.scuro.drsearch.task import Task
-from systemds.scuro.modality.modality import Modality
-from systemds.scuro.utils.checkpointing import CheckpointManager
-from systemds.scuro.utils.static_variables import DEBUG
+from systemds.scuro.representations.aggregated_representation import (
+    AggregatedRepresentation,
+)
+from systemds.scuro.utils.schema_helpers import get_shape
 
-# ----------------------------
-# Genome / Individual encoding
-# ----------------------------
-
-Tree = Any  # int leaf index OR tuple(left_subtree, right_subtree)
+Tree = int | Tuple["Tree", "Tree"]
 
 
-def genome_to_dag(genome) -> RepresentationDag:
-    builder = RepresentationDAGBuilder()
-    leaf_ids = [
-        builder.create_leaf_node(mod_id, repr_idx) for mod_id, repr_idx in genome.leaves
-    ]
-
-    def build(subtree: Tree, path: str) -> str:
-        if isinstance(subtree, int):
-            return leaf_ids[subtree]
-        left, right = subtree
-        left_id = build(left, path + "L")
-        right_id = build(right, path + "R")
-        op_cls = genome.fusion_ops[path]
-        op = op_cls()
-        return builder.create_operation_node(
-            op.__class__, [left_id, right_id], op.get_current_parameters()
-        )
-
-    return builder.build(build(genome.tree, ""))
-
-
-def _evaluate_individual_worker(
-    dag_pickle: bytes,
-    task_pickle: bytes,
-    modalities_pickle: bytes,
-    metric_name: str,
-) -> Tuple[float, Dict[str, Any]]:
-    dag = pickle.loads(dag_pickle)
-    task = pickle.loads(task_pickle)
-    modalities = pickle.loads(modalities_pickle)
-
-    start_time = time.time()
-    fused_representation = dag.execute(modalities, task)
-    scores = task.run(fused_representation.data)
-    runtime = time.time() - start_time
-    fitness = scores[1].average_scores[metric_name]
-
-    objective = {
-        "train_score": scores[0].average_scores,
-        "val_score": scores[1].average_scores,
-        "test_score": scores[2].average_scores,
-        "val": fitness,
-        "runtime": runtime,
-        "representation_time": runtime,
-        "task_time": 0.0,
-    }
-    return fitness, objective
-
-
-@dataclass
-class Individual:
-    # Leaves are concrete unimodal choices: (modality_id, representation_index)
-    leaves: List[Tuple[str, int]]
-    # Binary tree over leaf indices in `leaves`
-    tree: Tree
-    # Fusion op class per internal tree-path key, e.g. "L", "RLL", ...
-    fusion_ops: Dict[str, Any]
-    fitness: float = float("-inf")
-    objective: Dict[str, float] = field(default_factory=dict)
-    dag: Optional[RepresentationDag] = None
-
-
-# ----------------------------
-# Optional helper abstraction
-# ----------------------------
-
-
-class MutationOperator:
-    name: str = "base"
-
-    def __call__(
-        self,
-        individual: Individual,
-        rng: random.Random,
-        context: Dict[str, Any],
-    ) -> Individual:
-        raise NotImplementedError()
-
-
-def _collect_internal_paths(subtree: Tree, path: str = "") -> List[str]:
-    if isinstance(subtree, int):
+def _collect_internal_paths(tree: Tree, path: str = "") -> List[str]:
+    if isinstance(tree, int):
         return []
-    left, right = subtree
+    left, right = tree
     return (
         [path]
         + _collect_internal_paths(left, path + "L")
@@ -116,768 +61,707 @@ def _collect_internal_paths(subtree: Tree, path: str = "") -> List[str]:
     )
 
 
-def _get_subtree(subtree: Tree, target_path: str) -> Tree:
-    if target_path == "":
-        return copy.deepcopy(subtree)
-    if isinstance(subtree, int):
-        raise ValueError(f"Path '{target_path}' does not exist in leaf subtree")
-    left, right = subtree
-    if target_path[0] == "L":
-        return _get_subtree(left, target_path[1:])
-    return _get_subtree(right, target_path[1:])
+def _get_subtree(tree: Tree, path: str) -> Tree:
+    if not path:
+        return copy.deepcopy(tree)
+    left, right = tree
+    branch = left if path[0] == "L" else right
+    return _get_subtree(branch, path[1:])
 
 
-def _replace_subtree(subtree: Tree, target_path: str, replacement: Tree) -> Tree:
-    if target_path == "":
+def _replace_subtree(tree: Tree, path: str, replacement: Tree) -> Tree:
+    if not path:
         return copy.deepcopy(replacement)
-    if isinstance(subtree, int):
-        raise ValueError(f"Path '{target_path}' does not exist in leaf subtree")
-    left, right = subtree
-    if target_path[0] == "L":
-        return (
-            _replace_subtree(left, target_path[1:], replacement),
-            copy.deepcopy(right),
-        )
-    return (copy.deepcopy(left), _replace_subtree(right, target_path[1:], replacement))
+    left, right = tree
+    if path[0] == "L":
+        return _replace_subtree(left, path[1:], replacement), copy.deepcopy(right)
+    return copy.deepcopy(left), _replace_subtree(right, path[1:], replacement)
 
 
-def _collect_leaf_indices(subtree: Tree) -> List[int]:
-    if isinstance(subtree, int):
-        return [subtree]
-    left, right = subtree
-    return _collect_leaf_indices(left) + _collect_leaf_indices(right)
+def _collect_leaf_indices(tree: Tree) -> List[int]:
+    if isinstance(tree, int):
+        return [tree]
+    return _collect_leaf_indices(tree[0]) + _collect_leaf_indices(tree[1])
 
 
-def _sample_random_binary_tree_from_leaves(
-    leaf_indices: List[int], rng: random.Random
-) -> Tree:
-    nodes: List[Tree] = list(leaf_indices)
-    while len(nodes) > 1:
-        i, j = rng.sample(range(len(nodes)), 2)
-        a = nodes.pop(max(i, j))
-        b = nodes.pop(min(i, j))
-        nodes.append((a, b))
-    return nodes[0]
+def _remove_leaf_from_tree(tree: Tree, leaf: int) -> Optional[Tree]:
+    if isinstance(tree, int):
+        return None if tree == leaf else tree
+    left = _remove_leaf_from_tree(tree[0], leaf)
+    right = _remove_leaf_from_tree(tree[1], leaf)
+    if left is None:
+        return right
+    if right is None:
+        return left
+    return left, right
 
 
-def _reindex_tree(subtree: Tree, index_map: Dict[int, int]) -> Tree:
-    if isinstance(subtree, int):
-        return index_map[subtree]
-    left, right = subtree
-    return (_reindex_tree(left, index_map), _reindex_tree(right, index_map))
-
-
-def _remove_leaf_from_tree(subtree: Tree, leaf_idx: int) -> Optional[Tree]:
-    if isinstance(subtree, int):
-        if subtree == leaf_idx:
-            return None
-        return subtree
-
-    left, right = subtree
-    new_left = _remove_leaf_from_tree(left, leaf_idx)
-    new_right = _remove_leaf_from_tree(right, leaf_idx)
-
-    if new_left is None:
-        return new_right
-    if new_right is None:
-        return new_left
-    return (new_left, new_right)
-
-
-def _path_is_under_prefix(path: str, prefix: str) -> bool:
-    return prefix == "" or path == prefix or path.startswith(prefix)
+def _reindex_tree(tree: Tree, index_map: Dict[int, int]) -> Tree:
+    if isinstance(tree, int):
+        return index_map[tree]
+    return _reindex_tree(tree[0], index_map), _reindex_tree(tree[1], index_map)
 
 
 def _rebuild_fusion_ops(
-    subtree: Tree,
-    existing_fusion_ops: Dict[str, Any],
+    tree: Tree,
+    existing: Dict[str, type],
     rng: random.Random,
-    fusion_operators: List[Any],
+    operators: List[type],
     randomized_prefixes: Optional[List[str]] = None,
-) -> Dict[str, Any]:
-    fusion_ops: Dict[str, Any] = {}
-    randomized_prefixes = randomized_prefixes or []
-    for path in _collect_internal_paths(subtree):
-        if any(_path_is_under_prefix(path, prefix) for prefix in randomized_prefixes):
-            fusion_ops[path] = rng.choice(fusion_operators)
-        elif path in existing_fusion_ops:
-            fusion_ops[path] = existing_fusion_ops[path]
-        else:
-            fusion_ops[path] = rng.choice(fusion_operators)
-    return fusion_ops
+) -> Dict[str, type]:
+    prefixes = randomized_prefixes or []
+    rebuilt = {}
+    for path in _collect_internal_paths(tree):
+        randomized = any(not prefix or path.startswith(prefix) for prefix in prefixes)
+        rebuilt[path] = (
+            rng.choice(operators)
+            if randomized or path not in existing
+            else existing[path]
+        )
+    return rebuilt
 
 
-# ----------------------------
-# GA Optimizer skeleton
-# ----------------------------
+@dataclass
+class FusionSearchResult:
+    dag: RepresentationDag
+    train_score: dict
+    val_score: dict
+    test_score: dict
+    runtime: float = 0.0
+    task_time: float = 0.0
+    representation_time: float = 0.0
+    task_name: str = ""
+
+    val_fold_scores: dict = field(default_factory=dict)
+    train_fold_scores: dict = field(default_factory=dict)
+    test_fold_scores: dict = field(default_factory=dict)
+
+    task_timing: dict = field(default_factory=dict)
+
+    generation: int = -1
+    eval_index: int = -1
+    t_since_search_start_s: float = 0.0
+    t_eval_end_unix: float = 0.0
 
 
-class MultimodalGAOptimizer:
-    """
-    GA-based multimodal optimizer skeleton.
-    Keeps constructor inputs compatible with original MultimodalOptimizer.
-    """
+@dataclass
+class DagGenome:
+    leaves: List[Tuple[str, int]]
+    tree: Tree
+    fusion_ops: Dict[str, type]
 
+
+_TIMING_OBJECTIVES = {"runtime", "task_time", "representation_time"}
+
+ObjectiveSpec = Tuple[str, str]  # (name, "max" | "min")
+
+
+def _objective_value(
+    name: str, val_score: Dict[str, float], timing: Dict[str, float]
+) -> float:
+    if name in _TIMING_OBJECTIVES:
+        return timing[name]
+    return val_score[name]
+
+
+def _failure_fitness(objective_specs: List[ObjectiveSpec]) -> Tuple[float, ...]:
+    return tuple(
+        float("-inf") if direction == "max" else float("inf")
+        for _, direction in objective_specs
+    )
+
+
+def _evaluate_genome_body(
+    dag: RepresentationDag,
+    task: Task,
+    modalities: List[Any],
+    objective_specs: List[ObjectiveSpec],
+) -> Tuple[Optional[Tuple[float, ...]], Optional[Dict[str, Any]]]:
+    start = time.time()
+    fused = dag.execute(modalities, task, enable_cache=False)
+    if fused is None:
+        return None, None
+
+    if isinstance(fused, dict):
+        fused = fused[list(fused.keys())[-1]]
+
+    if task.expected_dim == 1 and get_shape(fused.metadata) > 1:
+        fused = AggregatedRepresentation().transform(fused)
+
+    t0 = time.time()
+    scores = task.run(fused.data)
+    task_time = time.time() - t0
+    total = time.time() - start
+
+    val_score = scores[1].average_scores
+    timing = {
+        "runtime": total,
+        "task_time": task_time,
+        "representation_time": total - task_time,
+    }
+    fitness = tuple(
+        _objective_value(name, val_score, timing) for name, _ in objective_specs
+    )
+    payload = {
+        "train_score": scores[0].average_scores,
+        "val_score": val_score,
+        "test_score": scores[2].average_scores,
+        "train_fold_scores": scores[0].fold_scores(),
+        "val_fold_scores": scores[1].fold_scores(),
+        "test_fold_scores": scores[2].fold_scores(),
+        "task_timing": getattr(task, "last_run_timing", {}),
+        **timing,
+    }
+    return fitness, payload
+
+
+def _evaluate_dag_worker(
+    dag_bytes: bytes,
+    task_bytes: bytes,
+    modalities_bytes: bytes,
+    objective_specs: List[ObjectiveSpec],
+) -> Tuple[Tuple[float, ...], Optional[Dict[str, Any]], Optional[str]]:
+    try:
+        dag = pickle.loads(dag_bytes)
+        task = pickle.loads(task_bytes)
+        modalities = pickle.loads(modalities_bytes)
+        fitness, payload = _evaluate_genome_body(dag, task, modalities, objective_specs)
+        if fitness is None:
+            fitness = _failure_fitness(objective_specs)
+        return fitness, payload, None
+    except Exception:
+        return _failure_fitness(objective_specs), None, traceback.format_exc()
+
+
+class MultimodalDeapOptimizer:
     def __init__(
         self,
         modalities: List[Any],
         unimodal_optimization_results: Any,
-        tasks: List[Any],
-        k: int = 2,
-        debug: bool = False,
+        tasks: List[Task],
+        debug: bool = True,
         min_modalities: int = 2,
         max_modalities: int = None,
         metric: str = "accuracy",
-        checkpoint_every: int = None,
-        resume: bool = True,
-        # --- GA controls (new) ---
+        objectives: Optional[List[ObjectiveSpec]] = None,
         population_size: int = 32,
         generations: int = 20,
-        elite_size: int = 4,
-        tournament_size: int = 3,
-        crossover_rate: float = 0.9,
-        mutation_rate: float = 0.3,
+        crossover_probability: float = 0.7,
+        mutation_probability: float = 0.4,
         random_seed: int = 42,
+        maximize_metric: bool = True,
+        elite_size: int = 2,
+        max_workers: int = 1,
+        batch_size: Optional[int] = None,
         early_stopping_patience: Optional[int] = 5,
         early_stopping_min_delta: float = 1e-6,
+        novelty_breeding: bool = True,
+        hall_of_fame_size: int = 5,
+        allow_repeated_modalities: bool = False,
     ):
         self.modalities = modalities
         self.tasks = tasks
-        self.k = k
         self.debug = debug
-        if DEBUG:
-            self.debug = True
-        self.metric_name = metric
-        self.min_modalities = max(2, min_modalities)
-        self.max_modalities = max_modalities or len(modalities)
+        self.allow_repeated_modalities = allow_repeated_modalities
 
-        self.population_size = population_size
-        self.generations = generations
-        self.elite_size = elite_size
-        self.tournament_size = tournament_size
-        self.crossover_rate = crossover_rate
-        self.mutation_rate = mutation_rate
-        self.random_seed = random_seed
-        self.rng = random.Random(random_seed)
-        self.early_stopping_patience = early_stopping_patience
-        self.early_stopping_min_delta = early_stopping_min_delta
+        self.min_modalities = max(1, min_modalities)
+        requested_max = max_modalities or len(modalities)
+        self.max_modalities = (
+            requested_max
+            if allow_repeated_modalities
+            else min(requested_max, len(modalities))
+        )
+        if self.max_modalities < self.min_modalities:
+            raise ValueError(
+                f"max_modalities ({self.max_modalities}) is below min_modalities "
+                f"({self.min_modalities})"
+            )
+        self.metric_name = metric
+        self.maximize_metric = maximize_metric
+
+        if objectives is not None:
+            if len(objectives) < 1:
+                raise ValueError(
+                    "objectives must contain at least one (name, direction) pair"
+                )
+            for name, direction in objectives:
+                if direction not in ("max", "min"):
+                    raise ValueError(
+                        f"objective direction must be 'max' or 'min', got "
+                        f"{direction!r} for objective {name!r}"
+                    )
+            self.objective_specs: List[ObjectiveSpec] = list(objectives)
+            self.metric_name = self.objective_specs[0][0]
+            self.maximize_metric = self.objective_specs[0][1] == "max"
+        else:
+            self.objective_specs = [
+                (self.metric_name, "max" if self.maximize_metric else "min")
+            ]
+        self.is_multi_objective = len(self.objective_specs) > 1
+
+        if len(self.modalities) < self.min_modalities:
+            raise ValueError(
+                f"MultimodalDeapOptimizer requires at least {self.min_modalities} "
+                f"modalities, got {len(self.modalities)}."
+            )
 
         self.operator_registry = Registry()
         self.fusion_operators = self.operator_registry.get_fusion_operators()
-
-        self.k_best_representations = self._extract_k_best_representations(
+        if not self.fusion_operators:
+            raise ValueError(
+                "MultimodalDeapOptimizer requires at least one registered "
+                "fusion operator."
+            )
+        self.k_best_representations = self._extract_k_best(
             unimodal_optimization_results
         )
 
-        self.optimization_results: Dict[str, List[OptimizationResult]] = {}
-        self._seen_results_by_task: Dict[str, Dict[str, OptimizationResult]] = {}
-        self._fitness_cache_by_task: Dict[
-            str, Dict[Tuple[Any, ...], Tuple[float, Dict[str, Any]]]
-        ] = {}
-        self._stats_by_task: Dict[str, Dict[str, int]] = {}
-        self._checkpoint_manager = CheckpointManager(
-            ".",
-            "multimodal_ga_checkpoint_",
-            checkpoint_every=checkpoint_every,
-            resume=resume,
+        self.optimization_results: Dict[str, List[FusionSearchResult]] = {}
+        self.evaluation_errors: Dict[str, int] = {}
+
+        self.hall_of_fame_size = max(1, hall_of_fame_size)
+        self.hall_of_fame: Dict[str, List[FusionSearchResult]] = {}
+        self._hof_fitness: Dict[str, List[Tuple[float, ...]]] = {}
+        self.rng = random.Random(random_seed)
+        self._eval_counter = 0
+        self._current_generation = -1
+        self._search_start = time.perf_counter()
+        self.population_size = max(1, population_size)
+        self.generations = max(1, generations)
+        self.crossover_probability = crossover_probability
+        self.mutation_probability = mutation_probability
+        self.random_seed = random_seed
+        self._fitness_cache: Dict[str, Dict[Tuple, Tuple[float, ...]]] = {}
+
+        self.elite_size = max(0, min(elite_size, self.population_size - 1))
+        self.max_workers = max(1, max_workers)
+        self.batch_size = max(1, batch_size or self.max_workers)
+        self.early_stopping_patience = early_stopping_patience
+        self.early_stopping_min_delta = early_stopping_min_delta
+        self.novelty_breeding = novelty_breeding
+        self._current_task_name = None
+
+        desired_weights = tuple(
+            1.0 if direction == "max" else -1.0 for _, direction in self.objective_specs
         )
-
-        # Register mutation ops you want enabled
-        self.mutation_ops: List[MutationOperator] = [
-            MutateFusionOperator(),
-            MutateRepresentationIndex(),
-            MutateTreeRotation(),
-            MutateSubtreeResample(),
-            # Optional global search mutations:
-            MutateAddModality(),
-            MutateDropModality(),
-        ]
-
-    # ----------------------------
-    # Public entrypoint
-    # ----------------------------
+        self._objective_weights = desired_weights
+        existing_weights = getattr(
+            getattr(creator, "FitnessMax", None), "weights", None
+        )
+        if existing_weights != desired_weights:
+            if hasattr(creator, "Individual"):
+                del creator.Individual
+            if hasattr(creator, "FitnessMax"):
+                del creator.FitnessMax
+            creator.create("FitnessMax", base.Fitness, weights=desired_weights)
+            creator.create("Individual", list, fitness=creator.FitnessMax)
+        elif not hasattr(creator, "Individual"):
+            creator.create("Individual", list, fitness=creator.FitnessMax)
 
     def optimize(
         self,
-        max_evaluations_per_task: Optional[int] = None,
-        # If you later wire this to NodeExecutor, add params like:
-        # use_node_executor: bool = True,
-        # max_workers: int = ...
-    ) -> Dict[str, List[OptimizationResult]]:
-        self.rng = random.Random(self.random_seed)
-        self._resume_if_available()
-        self._ensure_results_initialized()
-
+    ) -> Dict[str, List[FusionSearchResult]]:
         for task in self.tasks:
             task_name = task.model.name
-            if self.debug:
-                print(f"[GA] Task={task_name} initialization")
+            self._current_task_name = task_name
+            self.optimization_results.setdefault(task_name, [])
+            self.evaluation_errors.setdefault(task_name, 0)
 
-            # 1) Initialize population stochastically
-            population = self._initialize_population(task_name)
+            self._eval_counter = 0
+            self._search_start = time.perf_counter()
 
-            eval_budget = 0
-            best_seen: Optional[Individual] = None
-            no_improvement_gens = 0
+            population = self._build_initial_population(task_name)
+            best_ever = None
+            no_improve = 0
 
-            # 2) Evolution loop
             for gen in range(self.generations):
-                if self.debug:
-                    print(f"[GA] Task={task_name}, generation={gen}")
-
-                # Evaluate all individuals (or only unevaluated)
+                self._current_generation = gen
                 self._evaluate_population(population, task)
 
-                # Track best
-                gen_best = max(population, key=lambda ind: ind.fitness)
-                if best_seen is None or (
-                    gen_best.fitness > best_seen.fitness + self.early_stopping_min_delta
-                ):
-                    best_seen = copy.deepcopy(gen_best)
-                    no_improvement_gens = 0
+                if self.is_multi_objective:
+                    front = tools.sortNondominated(
+                        population, len(population), first_front_only=True
+                    )[0]
+                    front_signature = frozenset(
+                        self._genome_signature(ind[0]) for ind in front
+                    )
+                    if best_ever is None or front_signature != best_ever:
+                        best_ever = front_signature
+                        no_improve = 0
+                    else:
+                        no_improve += 1
+                    debug_msg = f"front_size={len(front)}"
                 else:
-                    no_improvement_gens += 1
-
-                # Optional budget stop
-                eval_budget += self._stats_by_task[task_name][
-                    "last_generation_executed"
-                ]
-                if (
-                    max_evaluations_per_task is not None
-                    and eval_budget >= max_evaluations_per_task
-                ):
-                    break
-                if (
-                    self.early_stopping_patience is not None
-                    and no_improvement_gens >= self.early_stopping_patience
-                ):
-                    if self.debug:
-                        print(
-                            f"[GA] Task={task_name} early stopping after "
-                            f"{no_improvement_gens} no-improvement generations"
-                        )
-                    break
-
-                # Selection + variation + replacement
-                next_population = self._elitism(population)
-                seen_genotypes = {
-                    self._individual_signature(ind) for ind in next_population
-                }
-                duplicate_retry_budget = self.population_size * 10
-                duplicate_retries = 0
-                while len(next_population) < self.population_size:
-                    p1 = self._tournament_select(population)
-                    p2 = self._tournament_select(population)
-
-                    if self.rng.random() < self.crossover_rate:
-                        c1, c2 = self._crossover(p1, p2)
+                    gen_best = max(population, key=lambda ind: ind.fitness.values[0])
+                    if (
+                        best_ever is None
+                        or gen_best.fitness.values[0]
+                        > best_ever.fitness.values[0] + self.early_stopping_min_delta
+                    ):
+                        best_ever = self._clone_individual(gen_best)
+                        no_improve = 0
                     else:
-                        c1, c2 = copy.deepcopy(p1), copy.deepcopy(p2)
+                        no_improve += 1
+                    debug_msg = f"best={gen_best.fitness.values[0]:.4f}"
 
-                    c1 = self._mutate(c1, task_name)
-                    c2 = self._mutate(c2, task_name)
-
-                    added_c1 = self._append_if_unique(
-                        next_population, c1, seen_genotypes, self.population_size
-                    )
-                    added_c2 = self._append_if_unique(
-                        next_population, c2, seen_genotypes, self.population_size
-                    )
-                    if not added_c1 and not added_c2:
-                        duplicate_retries += 1
-                    else:
-                        duplicate_retries = 0
-
-                    if duplicate_retries >= duplicate_retry_budget:
-                        break
-
-                while len(next_population) < self.population_size:
-                    immigrant = self._sample_random_individual(task_name)
-                    self._append_if_unique(
-                        next_population, immigrant, seen_genotypes, self.population_size
-                    )
-
-                population = next_population
-
-            self._checkpoint_manager.checkpoint_if_due(
-                self.optimization_results, "eval_count_by_task"
-            )
-
-            if self.debug and best_seen is not None:
-                print(f"[GA] Task={task_name}, best_fitness={best_seen.fitness:.6f}")
-
-        return self.optimization_results
-
-    def optimize_parallel(
-        self,
-        max_combinations: Optional[int] = None,
-        max_workers: int = 2,
-        batch_size: int = 8,
-    ) -> Dict[str, List[OptimizationResult]]:
-        self.rng = random.Random(self.random_seed)
-        self._resume_if_available()
-        self._ensure_results_initialized()
-
-        for task in self.tasks:
-            task_name = task.model.name
-            if self.debug:
-                print(f"[GA-P] Task={task_name} initialization")
-
-            population = self._initialize_population(task_name)
-            eval_budget = 0
-            best_seen: Optional[Individual] = None
-            no_improvement_gens = 0
-
-            for gen in range(self.generations):
                 if self.debug:
-                    print(f"[GA-P] Task={task_name}, generation={gen}")
+                    print(
+                        f"[GA] task={task_name} gen={gen} {debug_msg} "
+                        f"no_improve={no_improve} "
+                        f"errors={self.evaluation_errors.get(task_name, 0)}"
+                    )
 
-                self._evaluate_population_parallel(
-                    population,
-                    task,
-                    max_workers=max_workers,
-                    batch_size=max(1, batch_size),
-                )
-
-                gen_best = max(population, key=lambda ind: ind.fitness)
-                if best_seen is None or (
-                    gen_best.fitness > best_seen.fitness + self.early_stopping_min_delta
-                ):
-                    best_seen = copy.deepcopy(gen_best)
-                    no_improvement_gens = 0
-                else:
-                    no_improvement_gens += 1
-
-                eval_budget += self._stats_by_task[task_name][
-                    "last_generation_executed"
-                ]
-                if max_combinations is not None and eval_budget >= max_combinations:
-                    break
-                if (
+                stagnated = (
                     self.early_stopping_patience is not None
-                    and no_improvement_gens >= self.early_stopping_patience
-                ):
-                    if self.debug:
+                    and no_improve >= self.early_stopping_patience
+                )
+                if stagnated or gen == self.generations - 1:
+                    if self.debug and stagnated:
                         print(
-                            f"[GA-P] Task={task_name} early stopping after "
-                            f"{no_improvement_gens} no-improvement generations"
+                            f"[GA] task={task_name} early stopping after "
+                            f"{no_improve} generations without improvement"
                         )
                     break
 
-                next_population = self._elitism(population)
-                seen_genotypes = {
-                    self._individual_signature(ind) for ind in next_population
-                }
-                duplicate_retry_budget = self.population_size * 10
-                duplicate_retries = 0
-                while len(next_population) < self.population_size:
-                    p1 = self._tournament_select(population)
-                    p2 = self._tournament_select(population)
-
-                    if self.rng.random() < self.crossover_rate:
-                        c1, c2 = self._crossover(p1, p2)
-                    else:
-                        c1, c2 = copy.deepcopy(p1), copy.deepcopy(p2)
-
-                    c1 = self._mutate(c1, task_name)
-                    c2 = self._mutate(c2, task_name)
-
-                    added_c1 = self._append_if_unique(
-                        next_population, c1, seen_genotypes, self.population_size
-                    )
-                    added_c2 = self._append_if_unique(
-                        next_population, c2, seen_genotypes, self.population_size
-                    )
-                    if not added_c1 and not added_c2:
-                        duplicate_retries += 1
-                    else:
-                        duplicate_retries = 0
-
-                    if duplicate_retries >= duplicate_retry_budget:
-                        break
-
-                while len(next_population) < self.population_size:
-                    immigrant = self._sample_random_individual(task_name)
-                    self._append_if_unique(
-                        next_population, immigrant, seen_genotypes, self.population_size
-                    )
-
-                population = next_population
-
-            self._checkpoint_manager.checkpoint_if_due(
-                self.optimization_results, "eval_count_by_task"
-            )
-
-            if self.debug and best_seen is not None:
-                print(f"[GA-P] Task={task_name}, best_fitness={best_seen.fitness:.6f}")
+                population = self._next_generation(population, task_name, task)
 
         return self.optimization_results
 
-    # ----------------------------
-    # Initialization
-    # ----------------------------
+    def _make_individual(self, genome: DagGenome):
+        return creator.Individual([genome])
 
-    def _initialize_population(self, task_name: str) -> List[Individual]:
-        population = []
-        seen_signatures = set()
-        retry_budget = self.population_size * 10
-        retries = 0
-        for _ in range(self.population_size):
-            while retries < retry_budget:
-                candidate = self._sample_random_individual(task_name)
-                signature = self._individual_signature(candidate)
-                if signature not in seen_signatures:
-                    seen_signatures.add(signature)
-                    population.append(candidate)
-                    break
-                retries += 1
-            if retries >= retry_budget:
-                # Fall back to potentially duplicated random individuals rather than
-                # stalling initialization on very small search spaces.
-                population.append(self._sample_random_individual(task_name))
-        return population
-
-    def _sample_random_individual(self, task_name: str) -> Individual:
-        leaves = self._sample_leaf_set(task_name)
-        tree = self._sample_random_binary_tree(len(leaves))
-        fusion_ops = {}
-        self._assign_random_fusion_ops(tree, fusion_ops, path="")
-        return Individual(leaves=leaves, tree=tree, fusion_ops=fusion_ops)
-
-    def _sample_leaf_set(self, task_name: str) -> List[Tuple[str, int]]:
-        # Only sample from modalities that have at least one representation
-        # for the current task.
-        task_reps = self.k_best_representations.get(task_name, {})
-        available_modalities = [
-            m.modality_id
-            for m in self.modalities
-            if len(task_reps.get(m.modality_id, [])) > 0
-        ]
-
-        if len(available_modalities) < 2:
-            raise ValueError(
-                f"Need at least 2 modalities with non-empty representations for task "
-                f"'{task_name}', found {len(available_modalities)}."
-            )
-
-        # Clamp sampling range to available modalities.
-        lower = max(2, self.min_modalities)
-        upper = min(self.max_modalities, len(available_modalities))
-        if lower > upper:
-            lower = upper
-
-        r = self.rng.randint(lower, upper)
-        chosen_modalities = self.rng.sample(available_modalities, r)
-
-        # Sample one representation index from top-k for each chosen modality.
-        return [
-            (mod_id, self.rng.randrange(len(task_reps[mod_id])))
-            for mod_id in chosen_modalities
-        ]
-
-    def _sample_random_binary_tree(self, n_leaves: int) -> Tree:
-        nodes: List[Tree] = list(range(n_leaves))
-        while len(nodes) > 1:
-            i, j = self.rng.sample(range(len(nodes)), 2)
-            a = nodes.pop(max(i, j))
-            b = nodes.pop(min(i, j))
-            nodes.append((a, b))
-        return nodes[0]
-
-    def _assign_random_fusion_ops(
-        self, subtree: Tree, fusion_ops: Dict[str, Any], path: str
-    ):
-        if isinstance(subtree, int):
-            return
-        fusion_ops[path] = self.rng.choice(self.fusion_operators)
-        left, right = subtree
-        self._assign_random_fusion_ops(left, fusion_ops, path + "L")
-        self._assign_random_fusion_ops(right, fusion_ops, path + "R")
-
-    # ----------------------------
-    # Evaluation
-    # ----------------------------
-
-    def _evaluate_population(self, population: List[Individual], task: Task) -> None:
-        # Placeholder: currently serial + per-individual eval.
-        # You can batch DAGs and evaluate with NodeExecutor later.
-        task_name = task.model.name
-        cache = self._fitness_cache_by_task.setdefault(task_name, {})
-        evals_before = self._stats_by_task[task_name]["executed_evaluations"]
-        for ind in population:
-            # Already evaluated and unchanged.
-            if ind.fitness != float("-inf"):
-                continue
-
-            signature = self._individual_signature(ind)
-            cached = cache.get(signature)
-            if cached is not None:
-                self._stats_by_task[task_name]["cache_hits"] += 1
-                ind.fitness, ind.objective = cached[0], dict(cached[1])
-                continue
-
-            dag = self._individual_to_dag(ind)
-            ind.dag = dag
-            modalities = [
-                self.k_best_representations[task_name][mod_id][repr_idx]
-                for mod_id, repr_idx in ind.leaves
-            ]
-            self._stats_by_task[task_name]["executed_evaluations"] += 1
-            fitness, objective = self._evaluate_individual(dag, task, modalities)
-            ind.fitness = fitness
-            ind.objective = objective
-            cache[signature] = (fitness, dict(objective))
-            self._record_result(task_name, dag, objective)
-        self._stats_by_task[task_name]["last_generation_executed"] = (
-            self._stats_by_task[task_name]["executed_evaluations"] - evals_before
-        )
-
-    def _evaluate_population_parallel(
-        self,
-        population: List[Individual],
-        task: Task,
-        max_workers: int,
-        batch_size: int,
-    ) -> None:
-        ctx = mp.get_context("spawn")
-        task_name = task.model.name
-        cache = self._fitness_cache_by_task.setdefault(task_name, {})
-        evals_before = self._stats_by_task[task_name]["executed_evaluations"]
-        task_pickle = pickle.dumps(copy.deepcopy(task))
-        futures = {}
-        pending_followers: Dict[Tuple[Any, ...], List[Individual]] = {}
-
-        def _collect_ready(done_futures):
-            for done in done_futures:
-                ind, dag, signature = futures.pop(done)
-                fitness, objective = done.result()
-                ind.fitness = fitness
-                ind.objective = objective
-                cache[signature] = (fitness, dict(objective))
-                self._record_result(task_name, dag, objective)
-                for follower in pending_followers.pop(signature, []):
-                    follower.fitness = fitness
-                    follower.objective = dict(objective)
-
-        with ProcessPoolExecutor(max_workers=max_workers, mp_context=ctx) as executor:
-            for ind in population:
-                # Already evaluated and unchanged.
-                if ind.fitness != float("-inf"):
-                    continue
-
-                signature = self._individual_signature(ind)
-                cached = cache.get(signature)
-                if cached is not None:
-                    self._stats_by_task[task_name]["cache_hits"] += 1
-                    ind.fitness, ind.objective = cached[0], dict(cached[1])
-                    continue
-
-                if signature in pending_followers:
-                    pending_followers[signature].append(ind)
-                    self._stats_by_task[task_name]["pending_signature_hits"] += 1
-                    continue
-
-                dag = self._individual_to_dag(ind)
-                ind.dag = dag
-                modalities = [
-                    self.k_best_representations[task_name][mod_id][repr_idx]
-                    for mod_id, repr_idx in ind.leaves
-                ]
-
-                fut = executor.submit(
-                    _evaluate_individual_worker,
-                    pickle.dumps(dag),
-                    task_pickle,
-                    pickle.dumps(modalities),
-                    self.metric_name,
-                )
-                futures[fut] = (ind, dag, signature)
-                pending_followers[signature] = []
-                self._stats_by_task[task_name]["executed_evaluations"] += 1
-
-                if len(futures) >= batch_size:
-                    done, _ = wait(set(futures.keys()), return_when=FIRST_COMPLETED)
-                    _collect_ready(done)
-
-            if futures:
-                done, _ = wait(set(futures.keys()))
-                _collect_ready(done)
-        self._stats_by_task[task_name]["last_generation_executed"] = (
-            self._stats_by_task[task_name]["executed_evaluations"] - evals_before
-        )
-
-    def _evaluate_individual(
-        self, dag: RepresentationDag, task: Task, modalities: List[Modality]
-    ) -> Tuple[float, Dict[str, float]]:
-        start_time = time.time()
-        fused_representation = dag.execute(modalities, task)
-        task_start_time = time.time()
-        scores = task.run(fused_representation.data)
-        task_end_time = time.time()
-        runtime = time.time() - start_time
-        fitness = scores[1].average_scores[self.metric_name]
-        objective = {
-            "train_score": scores[0].average_scores,
-            "val_score": scores[1].average_scores,
-            "test_score": scores[2].average_scores,
-            "val": fitness,
-            "runtime": runtime,
-            "representation_time": fused_representation.transform_time,
-            "task_time": task_end_time - task_start_time,
-        }
-        return fitness, objective
-
-    def _record_result(
-        self, task_name: str, dag: RepresentationDag, objective: Dict[str, Any]
-    ) -> None:
-        key = str(dag.compute_full_node_signature(dag.root_node_id))
-        if key in self._seen_results_by_task[task_name]:
-            return
-        result = OptimizationResult(
-            dag=dag,
-            val_score=objective.get("val_score", {}),
-            train_score=objective.get("train_score", {}),
-            test_score=objective.get("test_score", {}),
-            runtime=objective.get("runtime", 0.0),
-            representation_time=objective.get("representation_time", 0.0),
-            task_time=objective.get("task_time", 0.0),
-            task_name=task_name,
-        )
-        self._seen_results_by_task[task_name][key] = result
-        self.optimization_results[task_name].append(result)
-
-    def get_task_stats(self, task_name: str) -> Dict[str, int]:
-        return dict(self._stats_by_task.get(task_name, {}))
-
-    def _individual_signature(self, ind: Individual) -> Tuple[Any, ...]:
-        def _normalize_tree(subtree: Tree) -> Any:
-            if isinstance(subtree, int):
-                return subtree
-            left, right = subtree
-            return (_normalize_tree(left), _normalize_tree(right))
-
-        leaves_sig = tuple(ind.leaves)
-        tree_sig = _normalize_tree(ind.tree)
-        ops_sig = tuple(
-            sorted((path, op.__name__) for path, op in ind.fusion_ops.items())
-        )
-        return leaves_sig, tree_sig, ops_sig
+    def _clone_individual(self, ind):
+        clone = creator.Individual([copy.deepcopy(ind[0])])
+        if ind.fitness.valid:
+            clone.fitness.values = ind.fitness.values
+        return clone
 
     def _append_if_unique(
         self,
-        population: List[Individual],
-        candidate: Individual,
-        seen_genotypes: set,
-        target_size: int,
+        population: List[Any],
+        genome: DagGenome,
+        seen_signatures: set,
     ) -> bool:
-        if len(population) >= target_size:
+        if len(population) >= self.population_size:
             return False
-        sig = self._individual_signature(candidate)
-        if sig in seen_genotypes:
+        sig = self._genome_signature(genome)
+        if sig in seen_signatures:
             return False
-        seen_genotypes.add(sig)
-        population.append(candidate)
+        seen_signatures.add(sig)
+        population.append(self._make_individual(genome))
         return True
 
-    # ----------------------------
-    # Selection / crossover / mutation
-    # ----------------------------
+    def _build_initial_population(self, task_name: str) -> List[Any]:
+        population: List[Any] = []
+        seen: set = set()
+        retry_budget = max(20, self.population_size * 10)
+        retries = 0
+        while len(population) < self.population_size and retries < retry_budget:
+            genome = self._random_genome(task_name)
+            if self._append_if_unique(population, genome, seen):
+                retries = 0
+            else:
+                retries += 1
+        while len(population) < self.population_size:
+            population.append(self._make_individual(self._random_genome(task_name)))
+        return population
 
-    def _elitism(self, population: List[Individual]) -> List[Individual]:
-        ranked = sorted(population, key=lambda x: x.fitness, reverse=True)
-        return [copy.deepcopy(ind) for ind in ranked[: self.elite_size]]
+    def _next_generation(
+        self, population: List[Any], task_name: str, task: Task
+    ) -> List[Any]:
+        if self.is_multi_objective:
+            offspring = self._breed_offspring(
+                population, task_name, seen=self._novelty_archive(task_name)
+            )
+            self._evaluate_population(offspring, task)
+            combined = list(population) + list(offspring)
+            return list(tools.selNSGA2(combined, self.population_size))
 
-    def _tournament_select(self, population: List[Individual]) -> Individual:
-        contestants = self.rng.sample(
-            population, k=min(self.tournament_size, len(population))
-        )
-        return copy.deepcopy(max(contestants, key=lambda x: x.fitness))
+        ranked = sorted(population, key=lambda ind: ind.fitness.values[0], reverse=True)
+        elite = [self._clone_individual(ind) for ind in ranked[: self.elite_size]]
+        seen = {self._genome_signature(ind[0]) for ind in elite}
+        seen |= self._novelty_archive(task_name)
+        return self._breed_offspring(population, task_name, initial=elite, seen=seen)
 
-    def _crossover(
-        self, p1: Individual, p2: Individual
-    ) -> Tuple[Individual, Individual]:
-        """
-        Skeleton crossover:
-        - Safe version assumes compatible leaves (same modality set/order).
-        - If incompatible, fallback to op-only crossover.
-        """
-        c1, c2 = copy.deepcopy(p1), copy.deepcopy(p2)
+    def _novelty_archive(self, task_name: str) -> set:
+        if not self.novelty_breeding:
+            return set()
+        return set(self._fitness_cache.get(task_name, {}))
 
-        if self._compatible_for_subtree_crossover(c1, c2):
-            c1_paths = _collect_internal_paths(c1.tree)
-            c2_paths = _collect_internal_paths(c2.tree)
-            if c1_paths and c2_paths:
-                path1 = self.rng.choice(c1_paths)
-                path2 = self.rng.choice(c2_paths)
-                subtree1 = _get_subtree(c1.tree, path1)
-                subtree2 = _get_subtree(c2.tree, path2)
-                c1.tree = _replace_subtree(c1.tree, path1, subtree2)
-                c2.tree = _replace_subtree(c2.tree, path2, subtree1)
-                c1.fusion_ops = _rebuild_fusion_ops(
-                    c1.tree,
-                    c1.fusion_ops,
-                    self.rng,
-                    self.fusion_operators,
-                    randomized_prefixes=[path1],
-                )
-                c2.fusion_ops = _rebuild_fusion_ops(
-                    c2.tree,
-                    c2.fusion_ops,
-                    self.rng,
-                    self.fusion_operators,
-                    randomized_prefixes=[path2],
-                )
+    def _breed_offspring(
+        self,
+        population: List[Any],
+        task_name: str,
+        initial: Optional[List[Any]] = None,
+        seen: Optional[set] = None,
+    ) -> List[Any]:
+        next_population = list(initial) if initial else []
+        seen = set(seen) if seen else set()
+
+        retry_budget = max(20, self.population_size * 10)
+        retries = 0
+        tournsize = max(1, min(3, len(population)))
+        while len(next_population) < self.population_size and retries < retry_budget:
+            p1, p2 = tools.selTournament(population, 2, tournsize=tournsize)
+
+            if self.rng.random() < self.crossover_probability:
+                g1, g2 = self._crossover_genomes(p1[0], p2[0])
+            else:
+                g1, g2 = copy.deepcopy(p1[0]), copy.deepcopy(p2[0])
+
+            if self.rng.random() < self.mutation_probability:
+                g1 = self._mutate_genome(g1, task_name)
+            if self.rng.random() < self.mutation_probability:
+                g2 = self._mutate_genome(g2, task_name)
+
+            added1 = self._append_if_unique(next_population, g1, seen)
+            added2 = self._append_if_unique(next_population, g2, seen)
+            retries = 0 if (added1 or added2) else retries + 1
+
+        while len(next_population) < self.population_size:
+            genome = self._random_genome(task_name)
+            if not self._append_if_unique(next_population, genome, seen):
+                next_population.append(self._make_individual(genome))
+
+        return next_population
+
+    def _evaluate_population(self, population: List[Any], task: Task) -> None:
+        to_evaluate = [ind for ind in population if not ind.fitness.valid]
+        if not to_evaluate:
+            return
+        if self.max_workers > 1 and len(to_evaluate) > 1:
+            self._evaluate_individuals_parallel(to_evaluate, task)
         else:
-            # op-only crossover: mix operator assignments
-            all_keys = set(c1.fusion_ops.keys()) | set(c2.fusion_ops.keys())
-            for k in all_keys:
-                if self.rng.random() < 0.5:
-                    if k in c2.fusion_ops:
-                        c1.fusion_ops[k] = c2.fusion_ops[k]
-                else:
-                    if k in c1.fusion_ops:
-                        c2.fusion_ops[k] = c1.fusion_ops[k]
+            for ind in to_evaluate:
+                fitness = self._evaluate_genome(ind[0], task)
+                ind.fitness.values = fitness
 
-        # invalidate stale eval
-        c1.fitness, c1.objective, c1.dag = float("-inf"), {}, None
-        c2.fitness, c2.objective, c2.dag = float("-inf"), {}, None
-        return c1, c2
+    def _evaluate_individuals_parallel(
+        self, individuals: List[Any], task: Task
+    ) -> None:
+        task_name = task.model.name
+        cache = self._fitness_cache.setdefault(task_name, {})
+        ctx = mp.get_context("spawn")
+        task_bytes = pickle.dumps(task)
+        futures: Dict[Any, Tuple[Any, RepresentationDag, Tuple]] = {}
+        pending_followers: Dict[Tuple, List[Any]] = {}
 
-    def _compatible_for_subtree_crossover(self, a: Individual, b: Individual) -> bool:
-        return [m for m, _ in a.leaves] == [m for m, _ in b.leaves] and len(
-            a.leaves
-        ) == len(b.leaves)
+        def _drain(done_futures):
+            for fut in done_futures:
+                ind, dag, sig = futures.pop(fut)
+                fitness, payload, error = fut.result()
+                ind.fitness.values = fitness
+                self._record_evaluation(task_name, dag, payload, error)
+                cache[sig] = fitness
+                for follower in pending_followers.pop(sig, []):
+                    follower.fitness.values = fitness
 
-    def _mutate(self, ind: Individual, task_name: str) -> Individual:
-        out = copy.deepcopy(ind)
-        if self.rng.random() >= self.mutation_rate:
-            return out
+        with ProcessPoolExecutor(
+            max_workers=self.max_workers, mp_context=ctx
+        ) as executor:
+            for ind in individuals:
+                genome = ind[0]
+                sig = self._genome_signature(genome)
+                cached = cache.get(sig)
+                if cached is not None:
+                    ind.fitness.values = cached
+                    continue
+                if sig in pending_followers:
+                    pending_followers[sig].append(ind)
+                    continue
 
-        op = self.rng.choice(self.mutation_ops)
-        context = {
-            "task_name": task_name,
-            "fusion_operators": self.fusion_operators,
-            "k_best_representations": self.k_best_representations,
-            "min_modalities": self.min_modalities,
-            "max_modalities": self.max_modalities,
-            "all_modalities": [m.modality_id for m in self.modalities],
+                dag = self._genome_to_dag(genome)
+                modalities = list(
+                    chain.from_iterable(self.k_best_representations[task_name].values())
+                )
+                fut = executor.submit(
+                    _evaluate_dag_worker,
+                    pickle.dumps(dag),
+                    task_bytes,
+                    pickle.dumps(modalities),
+                    self.objective_specs,
+                )
+                futures[fut] = (ind, dag, sig)
+                pending_followers[sig] = []
+
+                if len(futures) >= self.batch_size:
+                    done, _ = wait(set(futures.keys()), return_when=FIRST_COMPLETED)
+                    _drain(done)
+
+            if futures:
+                done, _ = wait(set(futures.keys()))
+                _drain(done)
+
+    def _record_evaluation(
+        self,
+        task_name: str,
+        dag: RepresentationDag,
+        payload: Optional[Dict[str, Any]],
+        error: Optional[str],
+    ) -> None:
+        self.optimization_results.setdefault(task_name, [])
+        if error is not None or payload is None:
+            self.evaluation_errors[task_name] = (
+                self.evaluation_errors.get(task_name, 0) + 1
+            )
+            if self.debug and error is not None:
+                last_line = error.strip().splitlines()[-1] if error.strip() else error
+                print(
+                    f"[GA] genome evaluation failed for task={task_name}: {last_line}"
+                )
+            return
+
+        result = FusionSearchResult(
+            dag=dag,
+            train_score=payload["train_score"],
+            val_score=payload["val_score"],
+            test_score=payload["test_score"],
+            train_fold_scores=payload.get("train_fold_scores", {}),
+            val_fold_scores=payload.get("val_fold_scores", {}),
+            test_fold_scores=payload.get("test_fold_scores", {}),
+            task_timing=payload.get("task_timing", {}),
+            runtime=payload["runtime"],
+            task_time=payload["task_time"],
+            representation_time=payload["representation_time"],
+            task_name=task_name,
+            generation=self._current_generation,
+            eval_index=self._eval_counter,
+            t_since_search_start_s=time.perf_counter() - self._search_start,
+            t_eval_end_unix=time.time(),
+        )
+        self.optimization_results.setdefault(task_name, []).append(result)
+        self._update_hall_of_fame(task_name, result)
+        self._eval_counter += 1
+
+    def _dominates(self, a: Tuple[float, ...], b: Tuple[float, ...]) -> bool:
+        """True if objective tuple `a` Pareto-dominates `b`, direction-aware."""
+        wa = [w * v for w, v in zip(self._objective_weights, a)]
+        wb = [w * v for w, v in zip(self._objective_weights, b)]
+        return all(x >= y for x, y in zip(wa, wb)) and any(
+            x > y for x, y in zip(wa, wb)
+        )
+
+    def _update_hall_of_fame(self, task_name: str, result: FusionSearchResult) -> None:
+        timing = {
+            "runtime": result.runtime,
+            "task_time": result.task_time,
+            "representation_time": result.representation_time,
         }
-        out = op(out, self.rng, context)
-        out.fitness, out.objective, out.dag = float("-inf"), {}, None
-        return out
+        try:
+            fitness = tuple(
+                _objective_value(name, result.val_score, timing)
+                for name, _ in self.objective_specs
+            )
+        except KeyError:
+            return
 
-    # ----------------------------
-    # Genome -> DAG
-    # ----------------------------
+        hof = self.hall_of_fame.setdefault(task_name, [])
+        fits = self._hof_fitness.setdefault(task_name, [])
 
-    def _individual_to_dag(self, ind: Individual) -> RepresentationDag:
+        if self.is_multi_objective:
+            if any(self._dominates(f, fitness) or f == fitness for f in fits):
+                return
+            keep = [i for i, f in enumerate(fits) if not self._dominates(fitness, f)]
+            self.hall_of_fame[task_name] = [hof[i] for i in keep] + [result]
+            self._hof_fitness[task_name] = [fits[i] for i in keep] + [fitness]
+            return
+
+        weight = self._objective_weights[0]
+        hof.append(result)
+        fits.append(fitness)
+        order = sorted(
+            range(len(fits)), key=lambda i: weight * fits[i][0], reverse=True
+        )
+        order = order[: self.hall_of_fame_size]
+        self.hall_of_fame[task_name] = [hof[i] for i in order]
+        self._hof_fitness[task_name] = [fits[i] for i in order]
+
+    def get_hall_of_fame(self, task_name: str) -> List[FusionSearchResult]:
+        return list(self.hall_of_fame.get(task_name, []))
+
+    def _extract_k_best(self, unimodal_results) -> Dict[str, Dict[str, List[Any]]]:
+        k_best = {}
+        for task in self.tasks:
+            name = task.model.name
+            k_best[name] = {}
+            for modality in self.modalities:
+                _, cached_data = unimodal_results.get_k_best_results(
+                    modality, task, self.metric_name
+                )
+                k_best[name][modality.modality_id] = cached_data
+        return k_best
+
+    def _available_modality_ids(self, task_name: str) -> List[Any]:
+        reps = self.k_best_representations[task_name]
+        return [
+            m.modality_id
+            for m in self.modalities
+            if len(reps.get(m.modality_id, [])) > 0
+        ]
+
+    def _leaf_capacity(self, task_name: str) -> int:
+        reps = self.k_best_representations[task_name]
+        ids = self._available_modality_ids(task_name)
+        if not self.allow_repeated_modalities:
+            return len(ids)
+        return sum(len(reps[mid]) for mid in ids)
+
+    def _random_genome(self, task_name: str) -> DagGenome:
+        reps = self.k_best_representations[task_name]
+        available_modality_ids = self._available_modality_ids(task_name)
+        capacity = self._leaf_capacity(task_name)
+        if capacity < self.min_modalities:
+            raise ValueError(
+                f"Need at least {self.min_modalities} distinct leaves for task "
+                f"'{task_name}', but only {capacity} are available across "
+                f"{len(available_modality_ids)} modalities."
+            )
+
+        upper = min(self.max_modalities, capacity)
+        lower = min(self.min_modalities, upper)
+        r = self.rng.randint(lower, upper)
+
+        if self.allow_repeated_modalities:
+            pool = [
+                (mid, idx)
+                for mid in available_modality_ids
+                for idx in range(len(reps[mid]))
+            ]
+            leaves = self.rng.sample(pool, r)
+        else:
+            chosen = self.rng.sample(available_modality_ids, r)
+            leaves = [(mid, self.rng.randrange(len(reps[mid]))) for mid in chosen]
+
+        tree = self._random_binary_tree(len(leaves))
+        fusion_ops = {}
+        self._assign_fusion_ops(tree, fusion_ops, "")
+        return DagGenome(leaves=leaves, tree=tree, fusion_ops=fusion_ops)
+
+    def _internal_paths(self, tree):
+        return _collect_internal_paths(tree)
+
+    def _random_binary_tree(self, n: int) -> Tree:
+        nodes: List[Tree] = list(range(n))
+        while len(nodes) > 1:
+            i, j = self.rng.sample(range(len(nodes)), 2)
+            a, b = nodes.pop(max(i, j)), nodes.pop(min(i, j))
+            nodes.append((a, b))
+        return nodes[0]
+
+    def _assign_fusion_ops(self, subtree: Tree, ops: Dict[str, Any], path: str) -> None:
+        if isinstance(subtree, int):
+            return
+        ops[path] = self.rng.choice(self.fusion_operators)
+        left, right = subtree
+        self._assign_fusion_ops(left, ops, path + "L")
+        self._assign_fusion_ops(right, ops, path + "R")
+
+    def _genome_to_dag(self, genome: DagGenome) -> RepresentationDag:
         builder = RepresentationDAGBuilder()
-        leaf_ids = []
-        for modality_id, repr_idx in ind.leaves:
-            # Leaves reference the cached transformed modality directly.
-            # We collapse the unimodal path from raw->representation here, so
-            # downstream execution only needs to run multimodal fusion nodes.
-            leaf_ids.append(builder.create_leaf_node(modality_id, repr_idx))
+        leaf_ids = [
+            builder.create_leaf_node(mod_id, repr_idx)
+            for mod_id, repr_idx in genome.leaves
+        ]
 
         def build(subtree: Tree, path: str) -> str:
             if isinstance(subtree, int):
@@ -885,351 +769,255 @@ class MultimodalGAOptimizer:
             left, right = subtree
             left_id = build(left, path + "L")
             right_id = build(right, path + "R")
-            op_cls = ind.fusion_ops[path]
+            op_cls = genome.fusion_ops[path]
             op = op_cls()
             return builder.create_operation_node(
                 op.__class__, [left_id, right_id], op.get_current_parameters()
             )
 
-        root_id = build(ind.tree, "")
-        dag = builder.build(root_id)
-        return self._collapse_cached_unimodal_nodes(dag)
+        return builder.build(build(genome.tree, ""))
 
-    def _collapse_cached_unimodal_nodes(
-        self, dag: RepresentationDag
-    ) -> RepresentationDag:
-        """
-        Remove unary unimodal nodes that originate at leaves.
+    def _genome_signature(self, g: DagGenome) -> Tuple:
+        def norm(t: Tree):
+            return t if isinstance(t, int) else (norm(t[0]), norm(t[1]))
 
-        In GA multimodal search, leaf inputs are already transformed unimodal
-        representations taken from the unimodal optimizer cache. Any unary chain
-        attached to those leaves is redundant and can be bypassed to ensure only
-        multimodal (fusion) operations are executed.
-        """
-        node_by_id = {node.node_id: copy.deepcopy(node) for node in dag.nodes}
-        if dag.root_node_id not in node_by_id:
-            return dag
-
-        changed = True
-        while changed:
-            changed = False
-            node_ids = list(node_by_id.keys())
-            for node_id in node_ids:
-                if node_id not in node_by_id:
-                    continue
-                node = node_by_id[node_id]
-                if len(node.inputs) != 1:
-                    continue
-
-                parent_id = node.inputs[0]
-                parent = node_by_id.get(parent_id)
-                if parent is None:
-                    continue
-                if parent.inputs:
-                    continue
-
-                # Bypass unary node by rewiring all consumers to the leaf input.
-                for consumer in node_by_id.values():
-                    consumer.inputs = [
-                        parent_id if input_id == node_id else input_id
-                        for input_id in consumer.inputs
-                    ]
-                if dag.root_node_id == node_id:
-                    dag.root_node_id = parent_id
-                del node_by_id[node_id]
-                changed = True
-
-        return RepresentationDag(
-            list(node_by_id.values()), dag.root_node_id, dag.dag_id
+        return (
+            tuple(g.leaves),
+            norm(g.tree),
+            tuple(sorted((p, c.__name__) for p, c in g.fusion_ops.items())),
         )
 
-    # ----------------------------
-    # Existing helper compatibility
-    # ----------------------------
+    def _evaluate_genome(self, genome: DagGenome, task: Task) -> Tuple[float, ...]:
+        task_name = task.model.name
+        sig = self._genome_signature(genome)
+        cache = self._fitness_cache.setdefault(task_name, {})
+        if sig in cache:
+            return cache[sig]
 
-    def _extract_k_best_representations(
-        self, unimodal_optimization_results: Any
-    ) -> Dict[str, Dict[str, List[Any]]]:
-        k_best = {}
-        for task in self.tasks:
-            task_name = task.model.name
-            k_best[task_name] = {}
-            for modality in self.modalities:
-                _, cached_data = unimodal_optimization_results.get_k_best_results(
-                    modality, task, self.metric_name
-                )
-                k_best[task_name][modality.modality_id] = cached_data
-        return k_best
-
-    def _resume_if_available(self) -> None:
-        loaded = self._checkpoint_manager.resume_from_checkpoint(
-            "eval_count_by_task",
-            lambda results: {
-                t.model.name: len(results.get(t.model.name, [])) for t in self.tasks
-            },
+        dag = self._genome_to_dag(genome)
+        modalities = list(
+            chain.from_iterable(self.k_best_representations[task_name].values())
         )
-        if loaded:
-            results, _, _ = loaded
-            self.optimization_results = results
 
-    def _ensure_results_initialized(self):
-        if not isinstance(self.optimization_results, dict):
-            self.optimization_results = {}
-        for task in self.tasks:
-            task_name = task.model.name
-            self.optimization_results.setdefault(task_name, [])
-            seen = self._seen_results_by_task.setdefault(task_name, {})
-            self._fitness_cache_by_task.setdefault(task_name, {})
-            self._stats_by_task.setdefault(
-                task_name,
-                {
-                    "executed_evaluations": 0,
-                    "last_generation_executed": 0,
-                    "cache_hits": 0,
-                    "pending_signature_hits": 0,
-                },
+        try:
+            fitness, payload = _evaluate_genome_body(
+                dag, task, modalities, self.objective_specs
             )
-            if self.optimization_results[task_name]:
-                for result in self.optimization_results[task_name]:
-                    if result.dag is None:
-                        continue
-                    key = str(
-                        result.dag.compute_full_node_signature(result.dag.root_node_id)
-                    )
-                    seen.setdefault(key, result)
+            error = None
+            if fitness is None:
+                fitness = _failure_fitness(self.objective_specs)
+        except Exception:
+            fitness = _failure_fitness(self.objective_specs)
+            payload, error = None, traceback.format_exc()
 
-    def _to_optimization_results(
-        self, individuals: List[Individual], task_name: str
-    ) -> List[OptimizationResult]:
-        out = []
-        for ind in individuals:
-            # Keep consistent with old output shape: OptimizationResult list per task
-            out.append(
-                OptimizationResult(
-                    dag=ind.dag,
-                    val_score={
-                        self.metric_name: ind.objective.get("val", float("-inf"))
-                    },
-                    train_score={},
-                    test_score={},
-                    runtime=ind.objective.get("runtime", 0.0),
-                    representation_time=ind.objective.get("representation_time", 0.0),
-                    task_time=ind.objective.get("task_time", 0.0),
-                    task_name=task_name,
+        self._record_evaluation(task_name, dag, payload, error)
+        cache[sig] = fitness
+        return fitness
+
+    def _crossover_genomes(
+        self, g1: DagGenome, g2: DagGenome
+    ) -> Tuple[DagGenome, DagGenome]:
+        c1, c2 = copy.deepcopy(g1), copy.deepcopy(g2)
+
+        if c1.leaves == c2.leaves:
+            paths1 = self._internal_paths(c1.tree)
+            paths2 = self._internal_paths(c2.tree)
+            if paths1 and paths2:
+                path1 = self.rng.choice(paths1)
+                path2 = self.rng.choice(paths2)
+                subtree1 = _get_subtree(c1.tree, path1)
+                subtree2 = _get_subtree(c2.tree, path2)
+                c1.tree = _replace_subtree(c1.tree, path1, subtree2)
+                c2.tree = _replace_subtree(c2.tree, path2, subtree1)
+                c1.fusion_ops = _rebuild_fusion_ops(
+                    c1.tree,
+                    {**c1.fusion_ops, **c2.fusion_ops},
+                    self.rng,
+                    self.fusion_operators,
+                    randomized_prefixes=[path1],
                 )
-            )
-        return out
+                c2.fusion_ops = _rebuild_fusion_ops(
+                    c2.tree,
+                    {**c2.fusion_ops, **c1.fusion_ops},
+                    self.rng,
+                    self.fusion_operators,
+                    randomized_prefixes=[path2],
+                )
+            return c1, c2
 
+        shared_paths = set(c1.fusion_ops) & set(c2.fusion_ops)
+        for path in shared_paths:
+            if self.rng.random() < 0.5:
+                c1.fusion_ops[path], c2.fusion_ops[path] = (
+                    c2.fusion_ops[path],
+                    c1.fusion_ops[path],
+                )
+        return c1, c2
 
-# ----------------------------
-# Mutation operators (outline)
-# ----------------------------
+    def _mutate_genome(self, g: DagGenome, task_name: str) -> DagGenome:
+        op = self.rng.choice(
+            [
+                self._mutate_change_fusion,
+                lambda gg: self._mutate_swap_leaf_repr(gg, task_name),
+                lambda gg: self.mutate_add_leaf(gg, task_name),
+                self.mutate_remove_leaf,
+                self.mutate_replace_subtree,
+            ]
+        )
+        return op(g)
 
+    def _mutate_change_fusion(self, g: DagGenome) -> DagGenome:
+        g = copy.deepcopy(g)
+        paths = [p for p in g.fusion_ops]
+        if not paths:
+            return g
+        path = self.rng.choice(paths)
+        choices = [op for op in self.fusion_operators if op != g.fusion_ops[path]]
+        if choices:
+            g.fusion_ops[path] = self.rng.choice(choices)
+        return g
 
-class MutateFusionOperator(MutationOperator):
-    """
-    Change one internal fusion op class.
-    """
+    def _mutate_swap_leaf_repr(self, g: DagGenome, task_name: str) -> DagGenome:
+        g = copy.deepcopy(g)
+        i = self.rng.randrange(len(g.leaves))
+        mod_id, current = g.leaves[i]
+        k = len(self.k_best_representations[task_name][mod_id])
+        if k <= 1:
+            return g
+        taken = {
+            idx for j, (mid, idx) in enumerate(g.leaves) if mid == mod_id and j != i
+        }
+        choices = [idx for idx in range(k) if idx != current and idx not in taken]
+        if not choices:
+            return g
+        g.leaves[i] = (mod_id, self.rng.choice(choices))
+        return g
 
-    name = "mutate_fusion_operator"
-
-    def __call__(
-        self, individual: Individual, rng: random.Random, context: Dict[str, Any]
-    ) -> Individual:
-        out = copy.deepcopy(individual)
-        if not out.fusion_ops:
-            return out
-        key = rng.choice(list(out.fusion_ops.keys()))
-        current = out.fusion_ops[key]
-        candidates = [op for op in context["fusion_operators"] if op != current]
-        if candidates:
-            out.fusion_ops[key] = rng.choice(candidates)
-        return out
-
-
-class MutateRepresentationIndex(MutationOperator):
-    """
-    Keep modality set fixed, change repr_idx for one modality leaf.
-    """
-
-    name = "mutate_representation_index"
-
-    def __call__(
-        self, individual: Individual, rng: random.Random, context: Dict[str, Any]
-    ) -> Individual:
-        out = copy.deepcopy(individual)
-        if not out.leaves:
-            return out
-        i = rng.randrange(len(out.leaves))
-        modality_id, cur_idx = out.leaves[i]
-        task_name = context["task_name"]
-        reps = context["k_best_representations"][task_name][modality_id]
-        if len(reps) <= 1:
-            return out
-        new_idx = rng.randrange(len(reps))
-        while new_idx == cur_idx:
-            new_idx = rng.randrange(len(reps))
-        out.leaves[i] = (modality_id, new_idx)
-        return out
-
-
-class MutateTreeRotation(MutationOperator):
-    """
-    Local reassociation (e.g., ((a,b),c) <-> (a,(b,c))) on a random eligible subtree.
-    """
-
-    name = "mutate_tree_rotation"
-
-    def __call__(
-        self, individual: Individual, rng: random.Random, context: Dict[str, Any]
-    ) -> Individual:
-        out = copy.deepcopy(individual)
-        candidates: List[Tuple[str, str]] = []
-        for path in _collect_internal_paths(out.tree):
-            subtree = _get_subtree(out.tree, path)
-            if isinstance(subtree, int):
-                continue
-            left, right = subtree
-            if not isinstance(left, int):
-                candidates.append((path, "left"))
-            if not isinstance(right, int):
-                candidates.append((path, "right"))
-
-        if not candidates:
-            return out
-
-        path, direction = rng.choice(candidates)
-        subtree = _get_subtree(out.tree, path)
-        left, right = subtree
-
-        if direction == "left":
-            left_left, left_right = left
-            rotated = (left_left, (left_right, right))
+    def mutate_add_leaf(self, g: DagGenome, task_name: str) -> DagGenome:
+        if len(g.leaves) >= self.max_modalities:
+            return g
+        g = copy.deepcopy(g)
+        reps = self.k_best_representations[task_name]
+        if self.allow_repeated_modalities:
+            existing_leaves = set(g.leaves)
+            candidates = [
+                (mid, idx)
+                for mid in self._available_modality_ids(task_name)
+                for idx in range(len(reps[mid]))
+                if (mid, idx) not in existing_leaves
+            ]
+            if not candidates:
+                return g
+            new_leaf = self.rng.choice(candidates)
         else:
-            right_left, right_right = right
-            rotated = ((left, right_left), right_right)
+            existing = {l[0] for l in g.leaves}
+            available = [
+                m.modality_id
+                for m in self.modalities
+                if m.modality_id not in existing
+                and len(reps.get(m.modality_id, [])) > 0
+            ]
+            if not available:
+                return g
+            mod_id = self.rng.choice(available)
+            new_leaf = (mod_id, self.rng.randrange(len(reps[mod_id])))
+        new_idx = len(g.leaves)
+        g.leaves.append(new_leaf)
+        if isinstance(g.tree, int):
+            g.tree = (g.tree, new_idx)
+            g.fusion_ops = {"": self.rng.choice(self.fusion_operators)}
+        else:
+            paths = self._internal_paths(g.tree)
+            path = self.rng.choice(paths)
+            sub = _get_subtree(g.tree, path)
+            g.tree = _replace_subtree(g.tree, path, (sub, new_idx))
+            g.fusion_ops = _rebuild_fusion_ops(
+                g.tree,
+                g.fusion_ops,
+                self.rng,
+                self.fusion_operators,
+                randomized_prefixes=[path],
+            )
+        return g
 
-        out.tree = _replace_subtree(out.tree, path, rotated)
-        out.fusion_ops = _rebuild_fusion_ops(
-            out.tree,
-            out.fusion_ops,
-            rng,
-            context["fusion_operators"],
-            randomized_prefixes=[path],
-        )
-        return out
-
-
-class MutateSubtreeResample(MutationOperator):
-    """
-    Pick a random internal subtree, keep the same leaves, but resample:
-    - subtree topology
-    - fusion operators inside subtree
-    """
-
-    name = "mutate_subtree_resample"
-
-    def __call__(
-        self, individual: Individual, rng: random.Random, context: Dict[str, Any]
-    ) -> Individual:
-        out = copy.deepcopy(individual)
-        internal_paths = _collect_internal_paths(out.tree)
-        if not internal_paths:
-            return out
-
-        path = rng.choice(internal_paths)
-        subtree = _get_subtree(out.tree, path)
-        leaf_indices = _collect_leaf_indices(subtree)
-        if len(leaf_indices) < 2:
-            return out
-
-        new_subtree = _sample_random_binary_tree_from_leaves(leaf_indices, rng)
-        out.tree = _replace_subtree(out.tree, path, new_subtree)
-        out.fusion_ops = _rebuild_fusion_ops(
-            out.tree,
-            out.fusion_ops,
-            rng,
-            context["fusion_operators"],
-            randomized_prefixes=[path],
-        )
-        return out
-
-
-class MutateAddModality(MutationOperator):
-    """
-    Optional global mutation:
-    Add one new modality (if below max_modalities), choose repr_idx, fuse with current root.
-    """
-
-    name = "mutate_add_modality"
-
-    def __call__(
-        self, individual: Individual, rng: random.Random, context: Dict[str, Any]
-    ) -> Individual:
-        out = copy.deepcopy(individual)
-        if len(out.leaves) >= context["max_modalities"]:
-            return out
-
-        existing = {m for m, _ in out.leaves}
-        candidates = [m for m in context["all_modalities"] if m not in existing]
-        if not candidates:
-            return out
-
-        m = rng.choice(candidates)
-        task_name = context["task_name"]
-        reps = context["k_best_representations"][task_name][m]
-        if len(reps) == 0:
-            return out
-
-        repr_idx = rng.randrange(len(reps))
-        new_leaf_idx = len(out.leaves)
-        out.leaves.append((m, repr_idx))
-
-        # Wrap old tree with new root fusion
-        old_tree = out.tree
-        old_fusion_ops = copy.deepcopy(out.fusion_ops)
-        out.tree = (old_tree, new_leaf_idx)
-        out.fusion_ops = {"": rng.choice(context["fusion_operators"])}
-        for path, op in old_fusion_ops.items():
-            out.fusion_ops["L" + path] = op
-        return out
-
-
-class MutateDropModality(MutationOperator):
-    """
-    Optional global mutation:
-    Remove one modality leaf (if above min_modalities) and collapse tree.
-    """
-
-    name = "mutate_drop_modality"
-
-    def __call__(
-        self, individual: Individual, rng: random.Random, context: Dict[str, Any]
-    ) -> Individual:
-        out = copy.deepcopy(individual)
-        if len(out.leaves) <= context["min_modalities"]:
-            return out
-
-        leaf_idx = rng.randrange(len(out.leaves))
-        new_tree = _remove_leaf_from_tree(out.tree, leaf_idx)
+    def mutate_remove_leaf(self, g: DagGenome) -> DagGenome:
+        if len(g.leaves) <= self.min_modalities:
+            return g
+        g = copy.deepcopy(g)
+        drop = self.rng.randrange(len(g.leaves))
+        new_tree = _remove_leaf_from_tree(g.tree, drop)
         if new_tree is None:
-            return out
-
-        remaining_leaves = [leaf for i, leaf in enumerate(out.leaves) if i != leaf_idx]
-        index_map = {}
-        next_idx = 0
-        for old_idx in range(len(out.leaves)):
-            if old_idx == leaf_idx:
-                continue
-            index_map[old_idx] = next_idx
-            next_idx += 1
-
-        out.leaves = remaining_leaves
-        out.tree = _reindex_tree(new_tree, index_map)
-        out.fusion_ops = _rebuild_fusion_ops(
-            out.tree,
-            out.fusion_ops,
-            rng,
-            context["fusion_operators"],
+            return g
+        keep = [i for i in range(len(g.leaves)) if i != drop]
+        index_map = {old: new for new, old in enumerate(keep)}
+        g.leaves = [g.leaves[i] for i in keep]
+        g.tree = _reindex_tree(new_tree, index_map)
+        g.fusion_ops = _rebuild_fusion_ops(
+            g.tree, g.fusion_ops, self.rng, self.fusion_operators
         )
-        return out
+        return g
+
+    def mutate_replace_subtree(self, g: DagGenome) -> DagGenome:
+        g = copy.deepcopy(g)
+        paths = self._internal_paths(g.tree)
+        if not paths:
+            return g
+        path = self.rng.choice(paths)
+        sub = _get_subtree(g.tree, path)
+        if isinstance(sub, int):
+            return g
+
+        if self.rng.random() < 0.5:
+            child = sub[0] if self.rng.random() < 0.5 else sub[1]
+            candidate = _replace_subtree(g.tree, path, child)
+            kept = sorted(set(_collect_leaf_indices(candidate)))
+            if len(kept) >= self.min_modalities:
+                index_map = {old: new for new, old in enumerate(kept)}
+                g.leaves = [g.leaves[i] for i in kept]
+                g.tree = _reindex_tree(candidate, index_map)
+                g.fusion_ops = _rebuild_fusion_ops(
+                    g.tree, {}, self.rng, self.fusion_operators
+                )
+                return g
+
+        leaf_idxs = _collect_leaf_indices(sub)
+        new_sub = self._random_binary_tree(len(leaf_idxs))
+        local_map = {i: leaf_idxs[i] for i in range(len(leaf_idxs))}
+        new_sub = _reindex_tree(new_sub, local_map)
+        g.tree = _replace_subtree(g.tree, path, new_sub)
+        g.fusion_ops = _rebuild_fusion_ops(
+            g.tree,
+            g.fusion_ops,
+            self.rng,
+            self.fusion_operators,
+            randomized_prefixes=[path],
+        )
+        return g
+
+    def store_results(self, file_name: str = None, overwrite: bool = False) -> str:
+        if file_name is None:
+            timestr = time.strftime("%Y%m%d-%H%M%S")
+            file_name = f"multimodal_optimizer_{timestr}.pkl"
+
+        directory = os.path.dirname(file_name) or "."
+        os.makedirs(directory, exist_ok=True)
+
+        if os.path.exists(file_name) and not overwrite:
+            raise FileExistsError(
+                f"Refusing to overwrite existing results file '{file_name}'. "
+                "Pass overwrite=True if this is intentional, or choose a "
+                "different file_name."
+            )
+
+        fd, tmp_path = tempfile.mkstemp(
+            dir=directory, prefix=".tmp_multimodal_results_", suffix=".pkl"
+        )
+        try:
+            with os.fdopen(fd, "wb") as f:
+                pickle.dump(self.optimization_results, f)
+            os.replace(tmp_path, file_name)
+        except Exception:
+            if os.path.exists(tmp_path):
+                os.remove(tmp_path)
+            raise
+        return file_name
